feat: split GPG-compat flags into a new tclig binary
tcli had two audiences: humans (import, list, describe, agent, card
status) and programs invoking it as `gpg.program` (git, pass). The
50-flag tcli --help mixed both. Split them:

tcli   -> human-facing key management + SSH agent (~18 flags)
tclig  -> GnuPG drop-in for git, pass, and anything with gpg.program
