+++ src/main.rs
@@ -1,10 +1,10 @@
 mod cli;
 
-use std::io::{stderr, stdin, stdout};
+use std::io::stdout;
 
 use clap::{CommandFactory, Parser};
 use clap_complete::generate;
-use tumpa_cli::{gpg, keystore, pinentry, ssh, store};
+use tumpa_cli::{keystore, pinentry, ssh, store};
 
 use cli::*;
 
@@ -15,46 +15,6 @@ fn main() {
     let keystore_path = args.keystore.clone();
 
     let res = match Mode::try_from(args) {
-        Ok(Mode::Sign { signer_id, armor }) => gpg::sign::sign(
-            stdin(),
-            stdout(),
-            stderr(),
-            &signer_id,
-            armor,
-            keystore_path.as_ref(),
-        ),
-        Ok(Mode::Verify { signature_file }) => gpg::verify::verify(
-            stdin(),
-            stdout(),
-            stderr(),
-            &signature_file,
-            keystore_path.as_ref(),
-        ),
-        Ok(Mode::Encrypt {
-            recipients,
-            output,
-            input,
-            armor,
-        }) => gpg::encrypt::encrypt(
-            input.as_ref(),
-            &output,
-            &recipients,
-            armor,
-            keystore_path.as_ref(),
-        ),
-        Ok(Mode::Decrypt { input, output }) => {
-            gpg::decrypt::decrypt(&input, output.as_ref(), keystore_path.as_ref())
-        }
-        Ok(Mode::DecryptListOnly { input }) => {
-            gpg::decrypt::decrypt_list_only(&input, keystore_path.as_ref())
-        }
-        Ok(Mode::ListKeysColon { key_ids }) => {
-            gpg::keys::list_keys_colon(&key_ids, keystore_path.as_ref())
-        }
-        Ok(Mode::ListSecretKeysColon) => {
-            gpg::keys::list_secret_keys_colon(keystore_path.as_ref())
-        }
-        Ok(Mode::ListConfig) => gpg::keys::list_config(),
         Ok(Mode::ListKeys) => list_keys(keystore_path.as_ref()),
         Ok(Mode::Agent {
             ssh,
@@ -133,22 +93,16 @@ fn print_help() {
         .unwrap_or_else(|| "tcli".to_string());
 
     eprintln!(
-        "tcli - GPG replacement and SSH agent backed by tumpa keystore
+                "tcli - key management and SSH agent backed by tumpa keystore
 
-To use with git:
+For git signing and pass integration, use `tclig` as the GPG drop-in:
 
-  $ git config --global gpg.program {exe}
-  $ git config --global user.signingkey <FINGERPRINT>
-  $ git config --global commit.gpgsign true
-
-To use with pass (password-store):
-
-  $ alias gpg={exe}
-  $ pass init <FINGERPRINT>
+    $ git config --global gpg.program tclig
+    $ ln -sf $(which tclig) ~/bin/gpg2
 
 To list keys in the keystore:
 
-  $ tcli --list-keys
+    $ {exe} --list-keys
 
 Keys are stored in ~/.tumpa/keys.db (managed by the tumpa desktop app).
 Hardware OpenPGP cards are tried first, then software keys from the keystore."
